+++ lib/src/awesome_notifications_core.dart
@@ -274,10 +274,14 @@ class AwesomeNotifications {
           content: pushNotification.content!,
           schedule: pushNotification.schedule,
           actionButtons: pushNotification.actionButtons);
-
     } catch (e) {
       return false;
     }
+  }
+
+  /// Opens the app notifications page
+  Future<void> showNotificationConfigPage() async {
+    await _channel.invokeMethod(CHANNEL_METHOD_SHOW_NOTIFICATION_PAGE);
   }
 
   /// Check if the notifications are permitted
@@ -422,36 +426,42 @@ class AwesomeNotifications {
 
   /// Dismiss all active notifications with the same channel key on status bar,
   /// without cancel the active respective schedules
-  Future<void>  dismissNotificationsByChannelKey(String channelKey) async {
-    await _channel.invokeMethod(CHANNEL_METHOD_DISMISS_NOTIFICATIONS_BY_CHANNEL_KEY, channelKey);
+  Future<void> dismissNotificationsByChannelKey(String channelKey) async {
+    await _channel.invokeMethod(
+        CHANNEL_METHOD_DISMISS_NOTIFICATIONS_BY_CHANNEL_KEY, channelKey);
   }
 
   /// Cancel all active schedules with the same channel key,
   /// without dismiss the respective notifications on status bar
-  Future<void>  cancelSchedulesByChannelKey(String channelKey) async {
-    await _channel.invokeMethod(CHANNEL_METHOD_CANCEL_SCHEDULES_BY_CHANNEL_KEY, channelKey);
+  Future<void> cancelSchedulesByChannelKey(String channelKey) async {
+    await _channel.invokeMethod(
+        CHANNEL_METHOD_CANCEL_SCHEDULES_BY_CHANNEL_KEY, channelKey);
   }
 
   /// Cancel and dismiss all notifications and schedules with the same channel key
-  Future<void>  cancelNotificationsByChannelKey(String channelKey) async {
-    await _channel.invokeMethod(CHANNEL_METHOD_CANCEL_NOTIFICATIONS_BY_CHANNEL_KEY, channelKey);
+  Future<void> cancelNotificationsByChannelKey(String channelKey) async {
+    await _channel.invokeMethod(
+        CHANNEL_METHOD_CANCEL_NOTIFICATIONS_BY_CHANNEL_KEY, channelKey);
   }
 
   /// Dismiss all active notifications with the same group key on status bar,
   /// without cancel the active respective schedules
-  Future<void>  dismissNotificationsByGroupKey(String groupKey) async {
-    await _channel.invokeMethod(CHANNEL_METHOD_DISMISS_NOTIFICATIONS_BY_GROUP_KEY, groupKey);
+  Future<void> dismissNotificationsByGroupKey(String groupKey) async {
+    await _channel.invokeMethod(
+        CHANNEL_METHOD_DISMISS_NOTIFICATIONS_BY_GROUP_KEY, groupKey);
   }
 
   /// Cancel all active schedules with the same group key,
   /// without dismiss the respective notifications on status bar
-  Future<void>  cancelSchedulesByGroupKey(String groupKey) async {
-    await _channel.invokeMethod(CHANNEL_METHOD_CANCEL_SCHEDULES_BY_GROUP_KEY, groupKey);
+  Future<void> cancelSchedulesByGroupKey(String groupKey) async {
+    await _channel.invokeMethod(
+        CHANNEL_METHOD_CANCEL_SCHEDULES_BY_GROUP_KEY, groupKey);
   }
 
   /// Cancel and dismiss all notifications and schedules with the same group key
-  Future<void>  cancelNotificationsByGroupKey(String groupKey) async {
-    await _channel.invokeMethod(CHANNEL_METHOD_CANCEL_NOTIFICATIONS_BY_GROUP_KEY, groupKey);
+  Future<void> cancelNotificationsByGroupKey(String groupKey) async {
+    await _channel.invokeMethod(
+        CHANNEL_METHOD_CANCEL_NOTIFICATIONS_BY_GROUP_KEY, groupKey);
   }
 
   /// Dismiss all active notifications, without cancel the active respective schedules